feat: make package autolinkable
Adapted the package to be linked by the autolinker properly. This change was partially tested. I tried to use this package in the RNOH tester app, but the app was failing on the ArkUINodeRegistry. The problem with the ArkUINodeRegistry should be fixed by https://gl.swmansion.com/rnoh/react-native-harmony-gesture-handler/-/issues/80. I could've avoided changing the names, but I did it on purpose to encourage end projects to use autolinking.

See merge request rnoh/react-native-harmony-gesture-handler!74