fix: prevent scroll locking when simultaneous NativeViewGestureHandlers are active
## Summary

When acquiring a scroll lock, we need to check if there are any simultaneous handlers that are NativeViewGestureHandlers. If so, we need to release the scroll lock. On Android when NativeViewGestureHandler is used, touch events are passed directly to the wrapped view's touch event handler. In the case of ScrollView, this behavior causes scrolling to be unlocked even when other handlers which should block scrolling are active.

## Test plan
Tested tester app for regressions.

See merge request rnoh/react-native-harmony-gesture-handler!101